backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
SpeechRecognition==3.10.4
pyttsx3==2.90
pyaudio==0.2.14
python-dotenv==1.0.0
APScheduler==3.10.4
requests==2.31.0
python-dateutil==2.8.2
psutil==5.9.6
supabase==2.3.0
PyJWT==2.8.0
bcrypt==4.1.2
groq==0.4.1
